refactor: Clean up playlist ETL file naming and remove version suffixes
This commit standardizes the playlist ETL file naming by:

1. **File Renames:**
   - `playlist_etl/transform.py` → `playlist_etl/transform_playlist.py` for clarity
   - `tests/transform2_test.py` → `tests/transform_test.py` (remove '2' suffix)
   - `tests/aggregate2_test.py` → `tests/aggregate_test.py` (remove '2' suffix)

2. **Import Updates:**
   - Updated `tests/integration_test.py` to import from `transform_playlist`
   - Updated `tests/transform_test.py` to import from `transform_playlist`
   - Verified `playlist_etl/main.py` already had correct import

3. **Documentation Updates:**
   - Updated `Makefile` line 63 to use `transform_playlist.py`
   - Updated `TESTING_GUIDE.md` references to use new test file names
   - Fixed all references to remove version suffixes from test files

4. **Verification:**
   - Confirmed all module imports work correctly
   - Ensured no broken references remain in codebase
   - Maintained backward compatibility for functionality

This refactoring improves code clarity by removing confusing version suffixes
and makes the transform module name more descriptive and consistent with
the codebase naming conventions.
